+++ presto-common/src/main/java/com/facebook/presto/common/array/LongBigArray.java
@@ -28,7 +28,7 @@ import static io.airlift.slice.SizeOf.sizeOfLongArray;
 // Copyright (C) 2010-2013 Sebastiano Vigna
 public final class LongBigArray
 {
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(LongBigArray.class).instanceSize();
+    private static final long INSTANCE_SIZE = ClassLayout.parseClass(LongBigArray.class).instanceSize();
     private static final long SIZE_OF_SEGMENT = sizeOfLongArray(SEGMENT_SIZE);
 
     private final long initialValue;
