+++ presto-common/src/main/java/com/facebook/presto/common/block/PageBuilderStatus.java
@@ -17,7 +17,7 @@ import org.openjdk.jol.info.ClassLayout;
 
 public class PageBuilderStatus
 {
-    public static final int INSTANCE_SIZE = ClassLayout.parseClass(PageBuilderStatus.class).instanceSize();
+    public static final long INSTANCE_SIZE = ClassLayout.parseClass(PageBuilderStatus.class).instanceSize();
     public static final int DEFAULT_MAX_PAGE_SIZE_IN_BYTES = 1024 * 1024;
 
     private final int maxPageSizeInBytes;
