+++ presto-main-base/src/main/java/com/facebook/presto/type/setdigest/SetDigest.java
@@ -53,8 +53,8 @@ public class SetDigest
     public static final int NUMBER_OF_BUCKETS = 2048;
     public static final int DEFAULT_MAX_HASHES = 8192;
     private static final int SIZE_OF_ENTRY = SIZE_OF_LONG + SIZE_OF_SHORT;
-    private static final int SIZE_OF_SETDIGEST = ClassLayout.parseClass(SetDigest.class).instanceSize();
-    private static final int SIZE_OF_RBTREEMAP = ClassLayout.parseClass(Long2ShortRBTreeMap.class).instanceSize();
+    private static final long SIZE_OF_SETDIGEST = ClassLayout.parseClass(SetDigest.class).instanceSize();
+    private static final long SIZE_OF_RBTREEMAP = ClassLayout.parseClass(Long2ShortRBTreeMap.class).instanceSize();
 
     private final HyperLogLog hll;
     private final Long2ShortSortedMap minhash;
@@ -129,7 +129,7 @@ public class SetDigest
         return hll;
     }
 
-    public int estimatedInMemorySize()
+    public long estimatedInMemorySize()
     {
         return hll.estimatedInMemorySize() + minhash.size() * SIZE_OF_ENTRY + SIZE_OF_SETDIGEST + SIZE_OF_RBTREEMAP;
     }
