+++ presto-orc/src/main/java/com/facebook/presto/orc/reader/DoubleBatchStreamReader.java
@@ -42,7 +42,7 @@ import static java.util.Objects.requireNonNull;
 public class DoubleBatchStreamReader
         implements BatchStreamReader
 {
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(DoubleBatchStreamReader.class).instanceSize();
+    private static final long INSTANCE_SIZE = ClassLayout.parseClass(DoubleBatchStreamReader.class).instanceSize();
 
     private final StreamDescriptor streamDescriptor;
