+++ presto-orc/src/main/java/com/facebook/presto/orc/reader/LongBatchStreamReader.java
@@ -37,7 +37,7 @@ import static java.util.Objects.requireNonNull;
 public class LongBatchStreamReader
         implements BatchStreamReader
 {
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(LongBatchStreamReader.class).instanceSize();
+    private static final long INSTANCE_SIZE = ClassLayout.parseClass(LongBatchStreamReader.class).instanceSize();
 
     private final StreamDescriptor streamDescriptor;
     private final LongDirectBatchStreamReader directReader;
