+++ presto-orc/src/main/java/com/facebook/presto/orc/stream/ByteOutputStream.java
@@ -34,7 +34,7 @@ import static java.lang.Math.toIntExact;
 public class ByteOutputStream
         implements ValueOutputStream<ByteStreamCheckpoint>
 {
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(ByteOutputStream.class).instanceSize();
+    private static final long INSTANCE_SIZE = ClassLayout.parseClass(ByteOutputStream.class).instanceSize();
 
     private static final int MIN_REPEAT_SIZE = 3;
     // A value outside of the range of a signed byte
