+++ presto-parquet/src/main/java/com/facebook/presto/parquet/DataPageV2.java
@@ -23,7 +23,7 @@ import static java.util.Objects.requireNonNull;
 public class DataPageV2
         extends DataPage
 {
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(DataPageV2.class).instanceSize();
+    private static final long INSTANCE_SIZE = ClassLayout.parseClass(DataPageV2.class).instanceSize();
 
     private final int rowCount;
     private final int nullCount;
